Use passive event listeners for window resize
By default, any event listener has the option to call
`e.preventDefault()` to prevent the default event from happening. This
means that the browser has to wait for the event handler to finish
before it can perform the default. For events like scrolling, this can
have a dramatically negative impact on the scrolling animation, which
needs to run at 60fps in order for it to feel good.

Thankfully, modern browsers offer us an escape hatch: passive event
listeners. By setting this flag, we are saying that `e.preventDefault()`
will never be called by the event handler. This makes it so the browser
is free to run the default.

In the case of a scroll event, this is the scroll animation. In the case
of window resize, I don't think this will have much of an effect. But, I
don't see any harm in doing it.