Build entry points
These three files are used as entry points into the project, and we
expect people to be importing them directly so they are at the top level
for convenience. Because of this, they have not been getting run through
Babel like the rest of the code here, so they were using older syntax.

In this commit, I am moving these files into the src directory so they
will be built into lib like the rest of the code here. In their place at
the top level, I am creating some very light shims that simply bring in
the built version of the module and re-export it if necessary.

This gives us a couple of advantages. First, it makes more of our code
consistent. Second, it makes it possible to produce two different kinds
of builds, one for CommonJS consumers and another for ES modules
consumers.